Use cargo-watch to run tests and rebuild
In an attempt to reduce outside tool usage, I changed from watchman to
cargo watch. It will automatically run rust and wasm tests when code
changes, then rebuild the wasm package if the tests run successfully.

I played with the console_error_panic_hook a bit, but was unable to get
it to work in my project so far. A panic still looks ugly for me, but
I got rid of the linter warning.

Found out I don't need to use cfg_if! macro like some of the Yew
templates do as #cfg works for single line statements just fine.