Read Remnant traits from file, parse CSV, use as default app entries
Not sure of what I'll do with app data yet, but having an accessible
file format sounds like a good idea. Added an id column so we can have
a unique column per item; may like to have it be a random string instead.

If we have something that's not just an id: u23, then we can manage data
upgrades for an item between app updates. Not sure that's needed here
though.