Remove TODO example actions not needed by app
Holy crap. Being able to delete items from an enum and watch it ripple
out to the rest of the app with red and yellow squiggles was very
satisfying. Removing one line showed me where app functionality should
be removed and where methods were no longer in use.

The use of macros (`html!`) did hide the cause of the error a bit as the
entire macro would turn red if any part of it would fail to build. The
errors were still helpful, it was just a bit more cognative overhead
to find the line.