Fix app to work offline
It needs full urls to serve from the serviceworker cache properly. Also,
code I copy/pasted used to cache a separate offline.html file specific
for running the offline version. It's not a bad idea, but for now I
use the same file (and assets) that the online version uses.