Add basic sharing api to app
At first I tried using web_sys, but I was having issues getting to the
`navigator` object and was confused if because Yew did not include the
"Navigator" feature, if I should see it at all. Even when I would
get a navigator object, it doesn't look like the share api methods are
available on it so I'd need to call js properties from Rust.

I tried adding web_sys myself to Cargo.toml and still had failures.
I eventually used the reflection API for js_web directly and could get
an instance of the navigation object, but it was getting messy. That led
me to the wasm-bindgen docs where I could create an interface to an
external javascript file which is WAY easier for me to move forward with.