Change from item index to explicit index to fix bug
Original example code used the items index in the entries Vec as the value
the Toggle message used to find the appropriate item in the list. When
I changed the code to allow filtering by name, the visible items indexes
did not map to the index they would have received unfiltered.

It's a better approach to not have to try and recreate filtering logic
in the toggle method. I'm not sure Box<String> is the right approach,
but it got me to where I needed to go.

I believe the callback closure is designed to run multiple times, and
if we capture the `entry` in the closure and call `id()` on it, then
we get into issues with lifetimes. If I make a local variable to the value
that `entry.id()` returns, then I can Box that each invocation and
dereference it during the comparision in the toggle method later on.